+++ src/components/myBlending/TeaCardList/TeaCardList.module.scss
@@ -1,0 +1,5 @@
+.cardList {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 2rem;
+}
